% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahn_letters.R
\name{ahn_letters}
\alias{ahn_letters}
\title{view AHN letter}
\usage{
ahn_letters()
}
\value{
data frame with ahn letters and their meaning
}
\description{
Show list of AHN letters to understand the naming scheme used by the PDOK to get the corresponding AHN dataset. The AHN letter is found  in the name of the AHN output files.
ahn_letters()
}
\details{
view AHN letters
}
\author{
Jelle Stuurman
}
