% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahn_point.R
\name{ahn_point}
\alias{ahn_point}
\title{Get AHN elevation of a certain point location}
\usage{
ahn_point(
  X,
  Y,
  name,
  LONLAT = FALSE,
  AHN = "AHN3",
  dem = "DSM",
  resolution = 0.5,
  destfile = "",
  decimals = 2,
  interpolate = TRUE,
  method.sheets = FALSE,
  keep.sheets = TRUE,
  redownload = FALSE
)
}
\arguments{
\item{X}{X coordinate in RD New or WGS84 (LON)}

\item{Y}{Y coordinate in RD New or WGS84 (LAT)}

\item{name}{Optional. Give a name of the specified area.}

\item{LONLAT}{Optional. Default FALSE. Set to TRUE if X and Y are in Longitude and Latitude format. Output will be in RD New format.}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{dem}{Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{resolution}{Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5, 25, and 100 m.}

\item{destfile}{Default "". All output point files be temporary files.When set to "structured", all the output point files will be saved in an organized way in the folder 'AHN_output' created in your current working directory. This included the AHN sheets. Set to any other output path to save all the output files there. The AHN sheets will be downloaded in the output folder 'AHN_sheets" in your working directory.}

\item{decimals}{Default 2. Decide number of decimal places of output elevations.}

\item{interpolate}{Default TRUE. Only applicable for AHN2 DTM. It decides if you want the interpolated version of the AHN2 or not.}

\item{method.sheets}{Default FALSE. FALSE downloads AHN area through the faster WCS method. Output is 32float geotif file. TRUE downloads AHN area through the available .tif AHN sheets (kaartbladen) available on PDOK.}

\item{keep.sheets}{Default TRUE. Only applicable if method.sheets is set to TRUE. Set to FALSE if you want to delete the downloaded sheets (kaartbladen).}

\item{redownload}{Default FALSE. Only applicable if sheets is set to TRUE. Set to TRUE if you want to redownload the sheets (kaartbladen)}
}
\value{
elevation in meters.
}
\description{
Get elevation of specific point location.
}
\details{
Get AHN elevation of a certain point location
}
\author{
Jelle Stuurman
}
