% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahn_points.R
\name{ahn_points}
\alias{ahn_points}
\title{Get AHN elvation of a certain point location}
\usage{
ahn_points(
  name,
  X,
  Y,
  LONLAT = FALSE,
  AHN = "AHN3",
  dem = "dsm",
  resolution = 0.5,
  delete.ahn = TRUE,
  decimals = 2,
  interpolate = TRUE
)
}
\arguments{
\item{name}{Optional. Give a name of the specified area.}

\item{X}{X coordidnate in RD New or WGS84 (LON)}

\item{Y}{Y coordidnate in RD New or WGS84 (LAT)}

\item{LONLAT}{Optional. Default FALSE. Set to TRUE if X and Y are in Longitude and Latitude format. Output will be in RD New format.}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{dem}{Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{resolution}{Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5, 25, and 100 m.}

\item{delete.ahn}{Deault TRUE. Set to FALSE if you want to keep the downloaded AHN area of the point (9 pixels).}

\item{decimals}{Default 2. Decide number of decimal places of output elevations.}

\item{interpolate}{Default TRUE. Ony applicable for AHN2 DTM. It decides if you want the interpolated version of the AHN2 or not.}
}
\value{
elevation in meters.
}
\description{
Get elevation of specific point location.
}
\details{
Get AHN elvation of a certain point location
}
\author{
Jelle Stuurman
ahn_points(name, X, Y, LONLAT = FALSE, AHN = "AHN3", dem = "dsm", resolution = 0.5, delete.ahn = TRUE, decimals = 2, interpolate = TRUE)
}
