% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahn_sheets_info.R
\name{ahn_sheets_info}
\alias{ahn_sheets_info}
\title{AHN sheets info}
\usage{
ahn_sheets_info(AHN, dem, resolution, output.dir)
}
\arguments{
\item{AHN}{Default 'AHN'. 'AHN' This is the latest full version made available by \href{https://www.pdok.nl/introductie/-/article/actueel-hoogtebestand-nederland-ahn}{PDOK} through their OGC Web services. Choose between 'AHN4', 'AHN3', 'AHN2', or 'AHN1'.}

\item{dem}{Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{resolution}{Default 0.5 meters for AHN2/AHN3/AHN4, 5 meters for AHN1. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5 and 100 m.}

\item{output.dir}{Optional but unnecessary. Set location of output raster files. Leaving blank (default) will make all output point files be temporary files. This output directory excludes the location of the AHN sheets which is depicted with the \code{sheets.dir} parameter.}
}
\description{
get polygon features that describe the AHN sheets
}
\author{
Jelle Stuurman
}
