% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_area.R
\name{create_area}
\alias{create_area}
\title{Create area}
\usage{
create_area(X, Y, radius, bbox, geom, LONLAT = FALSE, sheets = FALSE)
}
\arguments{
\item{X}{X coordidnate in RD New or WGS84 (LON)}

\item{Y}{Y coordidnate in RD New or WGS84 (LAT)}

\item{radius}{Optional. Set radius in meters of area around a point to create a buffer area.}

\item{bbox}{Optional. Set bbox of area. c(XMIN, YMIN, XMAX, YMAX)}

\item{geom}{Optional. Use geometric object as your area. .shp, .gpkg. Output will be BBox of this object.}

\item{LONLAT}{Optional. Default FALSE. Set to TRUE if X and Y are in Longitude and Latitude format. Output will be in RD New format.}

\item{sheets}{Default FALSE. Set to TRUE if you want to download AHN areas through the sheets (kaartbladen) instead through the WCS method (geotiff 32bit float}
}
\value{
"area": geometry of area, "bbox": BBOX coordinates.
}
\description{
Create area through buffer, BBOX or custom area
}
\details{
Get Create area
}
\author{
Jelle Stuurman
create_area(X, Y, radius, bbox, geom, LONLAT = FALSE, sheets = FALSE)
}
