% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bbox_polygon.R
\name{create_bbox_polygon}
\alias{create_bbox_polygon}
\title{create BBOX polygon}
\usage{
create_bbox_polygon(bbox)
}
\arguments{
\item{bbox}{Set bbox of area. c(XMIN, YMIN, XMAX, YMAX)}
}
\value{
BBOX polygon
}
\description{
Create BBOX geometry of certain area
}
\details{
Create BBOX polygon
}
\author{
Jelle Stuurman
}
