% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatialpoint.R
\name{create_spatialpoint}
\alias{create_spatialpoint}
\title{spatial point}
\usage{
create_spatialpoint(X, Y, LONLAT = FALSE)
}
\arguments{
\item{X}{X coordinate in RD New or WGS84 (LON)}

\item{Y}{Y coordinate in RD New or WGS84 (LAT)}

\item{LONLAT}{Optional. Default FALSE. Set to TRUE if X and Y are in Longitude and Latitude format. Output will be in RD New format.}
}
\value{
sf spatial point in RD New coordinates format
}
\description{
create a sf spatial point in RD New coordinates
}
\details{
Create a sf spatial point from X/Y or LAT/LON coordinates
}
\author{
Jelle Stuurman
create_spatialpoint(X = , Y = , LONLAT = TRUE)
}
