% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wcs_url.R
\name{create_wcs_url}
\alias{create_wcs_url}
\title{Create WCS URL}
\usage{
create_wcs_url(
  type,
  bbox,
  AHN = "AHN3",
  resolution = 0.5,
  dem = "DSM",
  interpolate = TRUE
)
}
\arguments{
\item{type}{required. Point or area.}

\item{bbox}{Optional. Set bbox of area. c(XMIN, YMIN, XMAX, YMAX)}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{resolution}{Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5 and 100 m.}

\item{dem}{Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{interpolate}{Default TRUE. Only applicable for AHN2 DTM. It decides if you want the interpolated version of the AHN2 or not.}
}
\value{
WCS URL string
}
\description{
Create WCS URL
create_wcs_url(type, bbox, AHN = "AHN3", resolution = 0.5, dem = "DSM", interpolate = TRUE)
}
\details{
Create WCS URL
}
\author{
Jelle Stuurman
}
