% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bladnrs.R
\name{download_bladnrs}
\alias{download_bladnrs}
\title{Download bladnrs}
\usage{
download_bladnrs(wd, AHN = "AHN3")
}
\arguments{
\item{wd}{Required. Working directory.}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}
}
\value{
all bladnrs geometry
}
\description{
Download blad numbers.
}
\details{
Download bladnrs
}
\author{
Jelle Stuurman
ahn_area(name, X, Y, radius, bbox, geom, LONLAT = FALSE, AHN = "AHN3", dem = "DSM", resolution, interpolate = TRUE, decimals = 2, sheets = FALSE, delete.sheets = FALSE, redownload = FALSE)
}
