% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dsm.R
\name{download_dsm}
\alias{download_dsm}
\title{Download DSM}
\usage{
download_dsm(
  name,
  wd,
  AHN = "AHN3",
  dem = "dsm",
  resolution = 0.5,
  bladnrs,
  area,
  interpolate = TRUE,
  delete.sheets = TRUE,
  redownload = FALSE
)
}
\arguments{
\item{name}{Optional. Give a name of the specified area.}

\item{wd}{Required. Working directory.}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{dem}{Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{resolution}{Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5, 25, and 100 m.}

\item{area}{Required area to be downloaded}

\item{interpolate}{Default TRUE. Olny applicable for AHN2 DTM. It decides if you want the interpolated version of the AHN2 or not.}

\item{delete.sheets}{Deault TRUE. Only applicable if sheets is set to TRUE. Set to FALSE if you want to keep the downloaded sheets (kaartbladen).}

\item{redownload}{Deafult FALSE. nly applicable if sheets is set to TRUE. Set to TRUE if you want to redownload the sheets (kaartbladen)}

\item{badnrs}{Required. Blad numbers.}
}
\value{
.tif of DSM AHN area
}
\description{
Download DSM
}
\details{
Download DSM
}
\author{
Jelle Stuurman
download_dsm(name, wd, AHN = "AHN3", dem = "dsm", resolution = 0.5, bladnrs, area, interpolate = TRUE, delete.sheets = TRUE, redownload = FALSE)
}
