% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dtm.R
\name{download_dtm}
\alias{download_dtm}
\title{Download DTM}
\source{
\url{https://www.pdok.nl/datasets}
}
\usage{
download_dtm(
  name,
  wd,
  AHN = "AHN3",
  dem = "DSM",
  resolution = 0.5,
  radius,
  bladnrs,
  area,
  interpolate = TRUE,
  keep.sheets = TRUE,
  redownload = FALSE
)
}
\arguments{
\item{name}{Optional. Give a name of the specified area.}

\item{wd}{Required. Working directory.}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{dem}{Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{resolution}{Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5, 25, and 100 m.}

\item{radius}{Radius of circle or squared BBOX in meters.}

\item{bladnrs}{Required. Blad numbers.}

\item{area}{Required area to be downloaded}

\item{interpolate}{Default TRUE. Only applicable for AHN2 DTM. It decides if you want the interpolated version of the AHN2 or not.}

\item{keep.sheets}{Default TRUE. Only applicable if method.sheets is set to TRUE. Set to FALSE if you want to delete the downloaded sheets (kaartbladen).}

\item{redownload}{Default FALSE. Only applicable if sheets is set to TRUE. Set to TRUE if you want to redownload the sheets (kaartbladen)}
}
\value{
.tif of DTM AHN area
}
\description{
Download DTM
download_dtm(name, wd, AHN = "AHN3", dem = "DSM", resolution = 0.5, bladnrs, area, interpolate = TRUE, keep.sheets = TRUE, redownload = FALSE)
}
\details{
Download DTM
}
\author{
Jelle Stuurman
}
