% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pointcloud.R
\name{download_point_cloud}
\alias{download_point_cloud}
\title{Download point cloud}
\source{
\url{https://www.pdok.nl/datasets}
download_point_cloud(name, wd, AHN = "AHN3", bladnrs, area, interpolate = TRUE, delete.sheets = TRUE, redownload = FALSE)
}
\usage{
download_point_cloud(
  name,
  wd,
  AHN = "AHN3",
  bladnrs,
  area,
  interpolate = TRUE,
  delete.sheets = TRUE,
  redownload = FALSE
)
}
\arguments{
\item{name}{Optional. Give a name of the specified area.}

\item{wd}{Required. Working directory.}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{bladnrs}{Required. Blad numbers.}

\item{area}{Required area to be downloaded}

\item{interpolate}{Default TRUE. Olny applicable for AHN2 DTM. It decides if you want the interpolated version of the AHN2 or not.}

\item{delete.sheets}{Deault TRUE. Only applicable if sheets is set to TRUE. Set to FALSE if you want to keep the downloaded sheets (kaartbladen).}

\item{redownload}{Deafult FALSE. nly applicable if sheets is set to TRUE. Set to TRUE if you want to redownload the sheets (kaartbladen)}
}
\value{
.tif of DSM AHN area
}
\description{
Download DSM
}
\details{
Download point cloud
}
\author{
Jelle Stuurman
}
