% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pointcloud.R
\name{download_pointCloud}
\alias{download_pointCloud}
\title{Download point cloud}
\source{
\url{https://www.pdok.nl/datasets}
}
\usage{
download_pointCloud(
  name,
  wd,
  AHN = "AHN3",
  bladnrs,
  area,
  bboxes,
  radius,
  gefilterd = TRUE,
  keep.sheets = TRUE,
  redownload = FALSE
)
}
\arguments{
\item{name}{Optional. Give a name of the specified area.}

\item{wd}{Required. Working directory.}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{bladnrs}{Required. Blad numbers.}

\item{area}{Required area to be downloaded}

\item{bboxes}{Optional. individual bboxes of cropped sheets}

\item{radius}{Radius of circle or squared BBOX in meters.}

\item{gefilterd}{Default TRUE. Only applicable for AHN1 or AHN2. It decides if you want the gefilterd point clouds or not.}

\item{keep.sheets}{Default TRUE. Only applicable if method.sheets is set to TRUE. Set to FALSE if you want to delete the downloaded sheets (kaartbladen).}

\item{redownload}{Default FALSE. Only applicable if sheets is set to TRUE. Set to TRUE if you want to redownload the sheets (kaartbladen)}
}
\value{
.tif of DSM AHN area
}
\description{
Download DSM
}
\details{
Download point cloud
}
\author{
Jelle Stuurman
}
