% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_elevation.R
\name{extract_elevation}
\alias{extract_elevation}
\title{extract elevation}
\usage{
extract_elevation(ras, point, extract.method = "bilinear")
}
\arguments{
\item{ras}{Required. Raster}

\item{point}{Required. spatial point}

\item{method}{'bilinear' (default) or 'simple'}
}
\value{
GeoTIFF file of AHN area
}
\description{
extract elevation at certain point
}
\details{
extract elevation raster
}
\author{
Jelle Stuurman
}
