% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ahn_point.R
\name{generate_ahn_point}
\alias{generate_ahn_point}
\title{Generate AHN area}
\usage{
generate_ahn_point(name = "", X, Y, LONLAT = FALSE, resolution)
}
\arguments{
\item{name}{Optional. Give a name of the specified area.}

\item{X}{X coordidnate in RD New or WGS84 (LON)}

\item{Y}{Y coordidnate in RD New or WGS84 (LAT)}

\item{LONLAT}{Optional. Default FALSE. Set to TRUE if X and Y are in Longitude and Latitude format. Output will be in RD New format.}

\item{resolution}{Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5, 25, and 100 m.}
}
\value{
"Name:", "point", "bbox"
}
\description{
Get AHN of a certain area
}
\details{
Generate AHN point
}
\author{
Jelle Stuurman
generate_ahn_point(name = "", X, Y, LONLAT = FALSE, resolution)
}
