% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lowres_point.R
\name{generate_lowres_ahn_points}
\alias{generate_lowres_ahn_points}
\title{Generate low resolution AHN}
\usage{
generate_lowres_ahn_points(n, resolution)
}
\arguments{
\item{n}{number}

\item{resolution}{Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5, 25, and 100 m.}
}
\value{
"min" "max" coordinates.
}
\description{
Get low resolution AHN
}
\details{
Generate low resolution AHN
}
\author{
Jelle Stuurman
generate_lowres_ahn_points(n, resolution)
}
