% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ahn_letter.R
\name{get_ahn_letter}
\alias{get_ahn_letter}
\title{Get AHN letter}
\usage{
get_ahn_letter(AHN, dem, resolution, interpolate, method, gefilterd)
}
\arguments{
\item{AHN}{Optional. Give a name of the specified area.}

\item{dem}{Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{resolution}{resolution}

\item{interpolate}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{method}{'raster' or 'pc' (point cloud)}

\item{gefilterd}{Only applicable for point clouds data. TRUE for 'gefilterd'. FALSE for uitgefilterd data.}
}
\value{
AHN letter string
}
\description{
Get AHN letter
}
\details{
Get AHN letter
}
\author{
Jelle Stuurman
}
