% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ahn_sheets.R
\name{get_ahn_sheets}
\alias{get_ahn_sheets}
\title{Get cropped AHN sheets}
\usage{
get_ahn_sheets(
  name,
  area,
  type = "",
  AHN = "AHN3",
  resolution = 0.5,
  dem = "DSM",
  radius,
  interpolate = TRUE,
  gefilterd = FALSE,
  sheets.redownload = FALSE,
  output.dir = "AHN_output",
  sheets.location,
  sheets.keep = TRUE
)
}
\arguments{
\item{name}{Optional. Give a name of the specified area.}

\item{area}{Required area to be downloaded}

\item{type}{'raster' or 'pc'.}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{resolution}{Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5 and 100 m.}

\item{dem}{Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{radius}{Radius of circle or squared BBOX in meters.}

\item{interpolate}{Default TRUE. Only applicable for AHN2 DTM. It decides if you want the interpolated version of the AHN2 or not.}

\item{gefilterd}{Default FALSE. Only applicable for AHN1 or AHN2 point cloud data. It decides if you want to download the 'gefilterd' point cloud data set or the 'uitgefilterd' data set.}

\item{sheets.redownload}{Default FALSE. Only applicable if sheets is set to TRUE. Set to TRUE if you want to redownload the sheets (structure)}

\item{output.dir}{destination file.}

\item{sheets.location}{Optional. Default is the 'AHN_sheets' directory in working directory. Set directory where all the AHN sheets are stored or will be stored. Always use the correct directory structure and capitalization within the selected directory. Example directory structure: 'AHN3/DSM' or 'AHN2/DTM' Only use extracted files in their original name after download.}

\item{sheets.keep}{Default TRUE. Only applicable if sheets.method is set to TRUE. Set to FALSE if you want to delete the downloaded sheets (structure).}
}
\value{
GeoTIFF AHN kaartblad cropped to area
}
\description{
Get cropped AHN sheets
get_ahn_sheets(name, area, AHN = "AHN3", resolution = 0.5, dem = "DSM", interpolate = TRUE, sheets.redownload = FALSE, sheets.keep = TRUE)
}
\details{
Get cropped AHN sheets
}
\author{
Jelle Stuurman
}
