% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resolution.R
\name{get_resolution}
\alias{get_resolution}
\title{Download DSM}
\usage{
get_resolution(AHN = "AHN3", resolution)
}
\arguments{
\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{resolution}{Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5, 25, and 100 m.}
}
\value{
"res": resolition in meters, "res_name:" name of resolution string.
}
\description{
Download DSM
}
\details{
Download DSM
}
\author{
Jelle Stuurman
get_resolution(AHN = "AHN3", resolution)
}
