% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surrounding_ahn.R
\name{get_surrounding_ahn}
\alias{get_surrounding_ahn}
\title{Get buffer surrounding AHN}
\usage{
get_surrounding_ahn(
  X,
  Y,
  LONLAT = FALSE,
  AHN = "AHN3",
  dem = dem,
  resolution = resolution,
  radius = 50
)
}
\arguments{
\item{X}{X coordinate in RD New or WGS84 (LON)}

\item{Y}{Y coordinate in RD New or WGS84 (LAT)}

\item{LONLAT}{Optional. Default FALSE. Set to TRUE if X and Y are in Longitude and Latitude format. Output will be in RD New format.}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{dem}{Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{resolution}{Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5, 25, and 100 m.}

\item{radius}{Optional. Set radius in meters of area around a point to create a buffer area.}
}
\value{
BBOX of buffer area
}
\description{
Get buffer surrounding AHN
ahn_area(name, X, Y, radius, bbox, geom, LONLAT = FALSE, AHN = "AHN3", dem = "DSM", resolution, interpolate = TRUE, decimals = 2, sheets = FALSE, delete.sheets = FALSE, redownload = FALSE)
}
\details{
Get buffer surrounding AHN
}
\author{
Jelle Stuurman
}
