% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_raster.R
\name{intersect_raster}
\alias{intersect_raster}
\title{AHn area}
\usage{
intersect_raster(ras, point)
}
\arguments{
\item{ras}{Required. Raster}

\item{point}{Required. spatial point}
}
\value{
.tif file of AHN area
}
\description{
Get AHN of a certain area
}
\details{
Get AHN of a certain area
}
\author{
Jelle Stuurman
ahn_area(name, X, Y, radius, bbox, geom, LONLAT = FALSE, AHN = "AHN3", dem = "dsm", resolution, interpolate = TRUE, decimals = 2, sheets = FALSE, delete.sheets = FALSE, redownload = FALSE)
}
