% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rAHNextract-package.R
\docType{package}
\name{rAHNextract-package}
\alias{rAHNextract}
\alias{rAHNextract-package}
\title{rAHNextract: Extract elevation points or areas in the Netherlands from the AHN}
\description{
Extract elevation points or areas in the Netherlands from the AHN.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Jellest/rAHNextract}
  \item Report bugs at \url{https://github.com/Jellest/rAHNextract/issues}
}

}
\keyword{internal}
