% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_name.R
\name{trim_name}
\alias{trim_name}
\title{Trim name}
\usage{
trim_name(name, name.supplement = "")
}
\arguments{
\item{name}{Name of AWS}

\item{name.supplement}{Optional. If an name.supplemental string needs to be added to AWS string name}
}
\value{
String with trimmed name
}
\description{
Trim name of AWS by removing spaces
}
\details{
Trim name of entry by removing spaces
}
\author{
Jelle Stuurman
}
