% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pc.R
\name{write_pc}
\alias{write_pc}
\title{write point cloud}
\source{
\url{https://www.pdok.nl/datasets}
}
\usage{
write_pc(temp_name, ahn_pc_name, nr, bladnrsLength, area)
}
\arguments{
\item{temp_name}{Required. Laz data}

\item{ahn_pc_name}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'}

\item{nr}{bladnr}

\item{bladnrsLength}{Total amunt of blad nrs}

\item{area}{Required area to be downloaded}
}
\description{
read point cloud
}
\details{
write point cloud
}
\author{
Jelle Stuurman
write_pc(temp_name, ahn_pc_name, nr, bladnrsLength, area)
}
