% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahn_area.R
\name{ahn_area}
\alias{ahn_area}
\title{AHN area}
\usage{
ahn_area(
  X,
  Y,
  radius,
  bbox,
  polygon,
  name,
  LONLAT = FALSE,
  AHN = "AHN3",
  dem = "DSM",
  type = "raster",
  resolution,
  interpolate = TRUE,
  gefilterd = FALSE,
  decimals = 2,
  destfile = "structured",
  method.sheets = FALSE,
  keep.sheets = TRUE,
  redownload = FALSE
)
}
\arguments{
\item{X}{Optional. X coordinate in RD New or WGS84 (LON)}

\item{Y}{Optional. Y coordinate in RD New or WGS84 (LAT)}

\item{radius}{Optional. Set radius in meters of area around a point to create a buffer area (circle).}

\item{bbox}{Optional. Set bbox of area. c(XMIN, YMIN, XMAX, YMAX) in RD New or WGS84 (LONLAT)}

\item{polygon}{Optional. Use spatial object as your area. .shp, .gpkg. Output will in the shape of this polygon}

\item{name}{Optional. Give a name of the specified area. Name will used in the gefilterd folders and files}

\item{LONLAT}{Optional. Default FALSE. Set to TRUE if X and Y are in Longitude and Latitude format. Output will always be in RD New format}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'}

\item{dem}{Required for raster datasets. Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{type}{Default 'raster'. Select 'raster' to get raster data}

\item{resolution}{Default 0.5 meters for AHN2/AHN3, 5 meters for AHN1. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5, and 100 m.}

\item{interpolate}{Default TRUE. Only applicable for AHN2 DTM. It decides if you want the interpolated version of the AHN2 or not.}

\item{gefilterd}{Default FALSE. Only applicable for AHN1 or AHN2 point cloud data. It decides if you want to download the 'gefilterd' point cloud data set or the 'uitgefilterd' data set.}

\item{decimals}{Default 2. Decide number of decimal places of gefilterd elevations.}

\item{destfile}{Default "structured". When set to "structured", all the gefilterd files will be saved in an organized way in the folder 'AHN_gefilterd' created in your current working directory. This included the AHN sheets. Set to 'Structured' to save  Set to any other gefilterd path to save all the gefilterd files there. The AHN sheets will be downloaded in the gefilterd folder 'AHN_sheets" in your working directory. Set to "", to have all gefilterd files be temporary files except for the AHN sheets which are downloaded in the 'AHN_sheets' folder.#'@param method.sheets Default FALSE. Set to TRUE if you want to download AHN areas through the sheets (kaartbladen) instead through the WCS method (geotif 32bit float}

\item{method.sheets}{Only applicable for raster data. Default FALSE. FALSE downloads AHN area through the faster WCS method. Output is 32float geotif file. TRUE downloads AHN area through the available .tiff AHN sheets (kaartbladen) available on PDOK.}

\item{keep.sheets}{Default TRUE. Only applicable if method.sheets is set to TRUE. Set to FALSE if you want to delete the downloaded sheets (kaartbladen).}

\item{redownload}{Default FALSE. Only applicable if sheets is set to TRUE. Set to TRUE if you want to redownload the sheets (kaartbladen)}
}
\value{
.tif file of AHN area
}
\description{
Get AHN of a certain area
}
\details{
Get AHN of a certain area
}
\author{
Jelle Stuurman
}
