% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahn_area.R
\name{ahn_area}
\alias{ahn_area}
\title{AHN raster area}
\usage{
ahn_area(
  name = "AHNarea",
  output.dir,
  X,
  Y,
  radius,
  bbox,
  polygon,
  AHN = "AHN3",
  dem = "DSM",
  resolution,
  interpolate = TRUE,
  LONLAT = FALSE,
  decimals = 2,
  sheets.method = FALSE,
  sheets.location,
  sheets.keep = TRUE,
  sheets.redownload = FALSE
)
}
\arguments{
\item{name}{Optional. Give a name of the specified area. This name will be used in the output file names.}

\item{output.dir}{Optional but recommended. Set location of output directory. Leaving blank (default) will make all output point files be temporary files. This output directory excludes the location of the AHN sheets which is depicted with the \code{sheets.location} parameter.}

\item{X}{Required for circle or BBOX when no BBOX coordinates are provided. X coordinate in RD New or WGS84 (LON)}

\item{Y}{Required for circle or BBOX when no BBOX coordinates are provided. Y coordinate in RD New or WGS84 (LAT)}

\item{radius}{Required for circle or BBOX when no BBOX coordinates are provided. Set radius in meters of area around a point to create a buffer area (circle).}

\item{bbox}{Required when using a BBOX. Create BBOX of an area. Use c(XMIN, YMIN, XMAX, YMAX) OR set to TRUE when a radius is provided. Use RD New or WGS84 (LONLAT) coordinates.}

\item{polygon}{Required when create an area of custom geometry. Use spatial object as your area. .shp, .gpkg. Output will in the shape of this polygon}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'}

\item{dem}{Required for raster datasets. Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{resolution}{Default 0.5 meters for AHN2/AHN3, 5 meters for AHN1. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5, and 100 m.}

\item{interpolate}{Default TRUE. Only applicable for AHN2 DTM. If true, it gets  the interpolated version of the AHN2.}

\item{LONLAT}{Optional. Default FALSE. Set to TRUE if X and Y are in Longitude and Latitude format. Output will always be in RD New format}

\item{decimals}{Default 2. Decide number of decimal places of output elevations.}

\item{sheets.method}{Default FALSE. FALSE downloads AHN area through the faster WCS method. Output is 32float GeoTIFF file.TRUE downloads AHN area through the available GeoTIFF AHN sheets available on \href{http://esrinl-content.maps.arcgis.com/apps/Embed/index.html?appid=a3dfa5a818174aa787392e461c80f781}{PDOK}.}

\item{sheets.location}{Optional. Default is the 'AHN_sheets' directory in the working directory. Set directory where all the AHN sheets are loaded when pre-existing sheets will be used or when new sheets will be stored. When loading existing files, always use the correct directory structure and capitalization within the selected directory. Example directory structure when this parameter is set to e.g. 'myFolder': 'myFolder/AHN_sheets/AHN3/DSM' or 'myFolder/AHN_sheets/AHN2/DTM'. Only use extracted files in their original name after download.}

\item{sheets.keep}{Default TRUE. Only applicable if \code{sheets.method} is set to TRUE and sheets were downloaded. Set to FALSE if you want to delete the downloaded sheet. It is recommended to keep the sheets if ahn elevation extarction will be followed.}
}
\value{
GeoTIFF file of AHN area
}
\description{
Get AHN raster area of a specified area.

AHN Area is retrieved through a circle (x, y and radius), BBOX (x, y and radius, OR bbox), or own geometry (polygon).

AHN data is obtained from the AHN3 (default), AHN2 or AHN 1 from the available resolutions.

Default resolution is always the highest resolution (smallest number). AHN3 and AHN2 DEM (Digital Elevation Model) are available for the Digital Surface Model (DSM) and Digital Terrain Model (DTM). AHN1 only has the DTM. THE DTM of the AHN2 has an interpolated and a non-interpolate version.

You can download the AHN data using the WCS method (default, \code{sheets.method = FALSE}) returning a GeoTIFF float32 format. The sheets method (\code{sheets.method = TRUE}) returns a regular raster GeoTIFF output file that is extracted from the PDOK sheets. The WCS method is recommended if only a few AHN elevation points need to be extracted. The sheets method always requires more data to be downloaded to the client but may be more efficient if many elevatiomns need to be retrieved from a small area. Choosing your method depends on speed and your desired output format. See documentation for all available parameters.
}
\author{
Jelle Stuurman
}
