% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_area.R
\name{create_area}
\alias{create_area}
\title{Create area}
\usage{
create_area(X, Y, radius, bbox, polygon, LONLAT = FALSE, type)
}
\arguments{
\item{X}{X coordinate in RD New or WGS84 (LON)}

\item{Y}{Y coordinate in RD New or WGS84 (LAT)}

\item{radius}{Optional. Set radius in meters of area around a point to create a buffer area.}

\item{bbox}{Optional. Set to TRUE if you want to create a bbox with the radius as parameter. Or create own bbox of area with format: c(XMIN, YMIN, XMAX, YMAX)}

\item{polygon}{Optional. Use polygon object as your area. Formats includes .shp, .gpkg. Output will be BBOX of this object.}

\item{LONLAT}{Optional. Default FALSE. Set to TRUE if X and Y are in Longitude and Latitude format. Output will be in RD New format.}

\item{type}{'point', 'raster' or 'pc'}
}
\value{
"area": polygon of area, "bbox": BBOX coordinates.
}
\description{
Create area through buffer, BBOX or custom area
create_area(X, Y, radius, bbox, polygon, LONLAT = FALSE, sheets = FALSE)
}
\details{
Get Create area
}
\author{
Jelle Stuurman
}
