% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pointcloud.R
\name{download_pointCloud}
\alias{download_pointCloud}
\title{Download point cloud}
\source{
\url{https://www.pdok.nl/datasets}
}
\usage{
download_pointCloud(
  name,
  output.dir,
  AHN = "AHN3",
  bladnrs,
  area,
  bboxes,
  radius,
  gefilterd = TRUE,
  sheets.location,
  sheets.keep = TRUE,
  sheets.redownload = FALSE
)
}
\arguments{
\item{name}{Optional. Give a name of the specified area.}

\item{output.dir}{Required. Working directory.}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{bladnrs}{Required. Blad numbers.}

\item{area}{Required area to be downloaded}

\item{bboxes}{Optional. individual bboxes of cropped sheets}

\item{radius}{Radius of circle or squared BBOX in meters.}

\item{gefilterd}{Default TRUE. Only applicable for AHN1 or AHN2. It decides if you want the gefilterd point clouds or not.}

\item{sheets.location}{Optional. Default is the 'AHN_sheets' directory in working directory. Set directory where all the AHN sheets are stored or will be stored. Always use the correct directory structure and capitalization within the selected directory. Example directory structure: 'AHN_sheets/AHN3/DSM' or 'AHN/sheets/AHN2/DTM'.  Always use the correct directory structure and capitalization within the selected directory when pre-existing sheets will be used.}

\item{sheets.keep}{Default TRUE. Only applicable if sheets.method is set to TRUE. Set to FALSE if you want to delete the downloaded sheets (structure).}

\item{sheets.redownload}{Default FALSE. Only applicable if sheets is set to TRUE. Set to TRUE if you want to redownload the sheets (structure)}
}
\value{
.laz of AHN area
}
\description{
Download DSM
}
\details{
Download point cloud
}
\author{
Jelle Stuurman
}
