% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wcs_raster.R
\name{download_wcs_raster}
\alias{download_wcs_raster}
\title{Download WCS raster}
\source{
\url{https://www.pdok.nl/datasets}
download_wcs_raster(name = "elevation", wcsUrl)
}
\usage{
download_wcs_raster(
  wcsUrl,
  name = "elevation",
  AHN = "AHN3",
  dem = "dsm",
  resolution,
  interpolate,
  keep.ahn
)
}
\arguments{
\item{wcsUrl}{Required. WCS URL}

\item{name}{Required. Name}

\item{AHN}{Default 'AHN3'. 'AHN3' 'AHN2' or 'AHN1'}

\item{dem}{Defailt 'dsm'. Selct DEM: 'dsm' or "dtm'}

\item{resolution}{Defailt 0.5}

\item{interpolate}{Only applicable for AHN2.}

\item{keep.ahn}{Set to FALSE if you want to delete the downloaded AHN area of the point (9 pixels).}
}
\value{
.tif float32 file of BBOX area.
}
\description{
Download WCS raster
}
\details{
Download WCS raster
}
\author{
Jelle Stuurman
}
