% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wcs_raster.R
\name{download_wcs_raster}
\alias{download_wcs_raster}
\title{Download WCS raster}
\source{
\url{https://www.pdok.nl/datasets}
download_wcs_raster(name = "elevation", wcsUrl)
}
\usage{
download_wcs_raster(wcsUrl, name = "elevation")
}
\arguments{
\item{wcsUrl}{Required. WCS URL}

\item{name}{Required. Name}
}
\value{
.tif float32 file of BBOX area.
}
\description{
Download WCS raster
}
\details{
Download WCS raster
}
\author{
Jelle Stuurman
}
