% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wcs_raster.R
\name{download_wcs_raster}
\alias{download_wcs_raster}
\title{Download WCS raster}
\source{
\url{https://www.pdok.nl/datasets}
download_wcs_raster(name = "elevation", wcsUrl)
}
\usage{
download_wcs_raster(
  wcsUrl,
  name = "elevation",
  AHN = "AHN3",
  dem = "DSM",
  resolution,
  radius,
  interpolate,
  destfile = "",
  type = "area"
)
}
\arguments{
\item{wcsUrl}{Required. WCS URL}

\item{name}{Required. Name}

\item{AHN}{Default 'AHN3'. 'AHN3' 'AHN2' or 'AHN1'}

\item{dem}{Default 'DSM'. Select DEM: 'DSM' or "DTM'}

\item{resolution}{Default 0.5}

\item{radius}{Radius of circle or squared BBOX in meters.}

\item{interpolate}{Only applicable for AHN2.}

\item{destfile}{destination file.}

\item{type}{'area' or 'point'}
}
\value{
.tif float32 file of BBOX area.
}
\description{
Download WCS raster
}
\details{
Download WCS raster
}
\author{
Jelle Stuurman
}
