% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_elevation.R
\name{extract_elevation}
\alias{extract_elevation}
\title{extract elevation}
\usage{
extract_elevation(ras, point, extract.method = "bilinear")
}
\arguments{
\item{ras}{Required. Raster}

\item{point}{Required. spatial point}

\item{extract.method}{Default 'bilinear'. Choose 'bilinear or 'simple'. Intersection is done using \href{https://www.rdocumentation.org/packages/raster/versions/3.1-5/topics/extract}{\code{extract()}} function from the \code{raster} package.}
}
\value{
GeoTIFF file of AHN area
}
\description{
extract elevation at certain point
}
\details{
extract elevation raster
}
\author{
Jelle Stuurman
}
