% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ahn_sheets.R
\name{get_ahn_sheets}
\alias{get_ahn_sheets}
\title{Get cropped AHN sheets}
\usage{
get_ahn_sheets(
  name,
  area,
  type = "",
  AHN = "AHN3",
  resolution = 0.5,
  dem = "dsm",
  interpolate = TRUE,
  filtered = FALSE,
  redownload = FALSE,
  keep.sheets = TRUE
)
}
\arguments{
\item{name}{Optional. Give a name of the specified area.}

\item{area}{Required area to be downloaded}

\item{type}{'raster' or 'pc'.}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.}

\item{resolution}{Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5, 25, and 100 m.}

\item{dem}{Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.}

\item{interpolate}{Default TRUE. Olny applicable for AHN2 DTM. It decides if you want the interpolated version of the AHN2 or not.}

\item{filtered}{Default FALSE. Only applicable for AHN1 or AHN2 point cloud data. It decides if you want to download the 'gefiltered' point cloud data set or the 'uitgefiltered' data set.}

\item{redownload}{Deafult FALSE. nly applicable if sheets is set to TRUE. Set to TRUE if you want to redownload the sheets (kaartbladen)}

\item{keep.sheets}{Default TRUE. Only applicable if method.sheets is set to TRUE. Set to FALSE if you want to delete the downloaded sheets (kaartbladen).}
}
\value{
.tif AHN kaartblad cropped to area
}
\description{
Get cropped AHN sheets
}
\details{
Get cropped AHN sheets
}
\author{
Jelle Stuurman
get_ahn_sheets(name, area, AHN = "AHN3", resolution = 0.5, dem = "dsm", interpolate = TRUE, redownload = FALSE, keep.sheets = TRUE)
}
