% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_raster.R
\name{intersect_raster}
\alias{intersect_raster}
\title{intersect raster}
\usage{
intersect_raster(ras, point, method = "bilinear")
}
\arguments{
\item{ras}{Required. Raster}

\item{point}{Required. spatial point}

\item{method}{'bilinear' (default) or 'simple'}
}
\value{
GeoTIFF file of AHN area
}
\description{
intersect raster at certain point
}
\details{
intersect raster
}
\author{
Jelle Stuurman
}
