% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pc.R
\name{read_pc}
\alias{read_pc}
\title{read point cloud}
\source{
\url{https://www.pdok.nl/datasets}
}
\usage{
read_pc(
  output.dir,
  laz,
  AHN = "AHN3",
  ahn_letter,
  area,
  radius,
  bladnrs,
  bbox,
  name,
  nr,
  bladnrsLength
)
}
\arguments{
\item{output.dir}{Required. working directory}

\item{laz}{Required. .laz data}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'}

\item{ahn_letter}{Code name of (uit)gefilterd data.}

\item{area}{Required area to be downloaded}

\item{radius}{Required for circle or squared bbox. Radius of selected AHN area.}

\item{bladnrs}{Required. Blad numbers.}

\item{bbox}{Optional. individual bboxes of cropped sheets}

\item{name}{Optional. Give a name of the specified area.}

\item{nr}{bladnr}

\item{bladnrsLength}{Total amount of blad nrs}
}
\value{
.laz data
}
\description{
read point cloud
}
\details{
read point cloud
}
\author{
Jelle Stuurman
}
