% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pc.R
\name{read_pc}
\alias{read_pc}
\title{read point cloud}
\source{
\url{https://www.pdok.nl/datasets}
read_pc(laz, AHN = "AHN3", filtered_name, area, radius, bladnrs, bbox, name, nr, bladnrsLength)
}
\usage{
read_pc(
  laz,
  AHN = "AHN3",
  filtered_name,
  area,
  radius,
  bladnrs,
  bbox,
  name,
  nr,
  bladnrsLength
)
}
\arguments{
\item{laz}{Required. Laz data}

\item{AHN}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'}

\item{filtered_name}{Code name of (un)filtered data.}

\item{area}{Required area to be downloaded}

\item{radius}{Optional. Set radius in meters of area around a point to create a buffer area (circle).}

\item{bladnrs}{Required. Blad numbers.}

\item{bbox}{Optional. individual bboxes of cropped sheets}

\item{name}{Optional. Give a name of the specified area.}

\item{nr}{bladnr}

\item{bladnrsLength}{Total amunt of blad nrs}
}
\value{
.laz data
}
\description{
read point cloud
}
\details{
read point cloud
}
\author{
Jelle Stuurman
}
