% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pc.R
\name{write_pc}
\alias{write_pc}
\title{write point cloud}
\usage{
write_pc(temp_name, name, ahn_pc_filename, nr, bladnrsLength, area, radius)
}
\arguments{
\item{temp_name}{Required. .laz data}

\item{name}{name of file}

\item{ahn_pc_filename}{Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'}

\item{nr}{bladnr}

\item{bladnrsLength}{Total amount of blad nrs}

\item{area}{Required area to be downloaded}

\item{radius}{radius. Required for correct naming.}
}
\description{
read point cloud
}
\details{
write point cloud
}
\author{
Jelle Stuurman
write_pc(temp_name, ahn_pc_filename, nr, bladnrsLength, area)
}
